+++ src/fastabx/subsample.py
@@ -23,7 +23,7 @@ def subsample_each_cell(df: pl.LazyFrame, size: int, seed: int) -> pl.LazyFrame:
     on per-cell independence of the retained items.
     """
     return (
-        df.with_columns(pl.concat_str(~INDEX_COLUMNS, separator="-").alias("__group"))
+        df.with_row_index("__group")
         .with_columns(
             INDEX_COLUMNS.explode(empty_as_null=False).shuffle(seed=seed).implode().over("__group").list.head(size)
         )
@@ -32,16 +32,15 @@ def subsample_each_cell(df: pl.LazyFrame, size: int, seed: int) -> pl.LazyFrame:
 
 
 def subsample_across_group(df: pl.LazyFrame, size: int, seed: int) -> pl.LazyFrame:
-    """Subsample each group of 'across' condition by taking ``size`` possible values for X in each group."""
+    """Retain up to ``size`` observed tuples of X conditions for each A/B condition group."""
     x_cols = [c for c in df.collect_schema() if c.endswith("_x") and c != "index_x"]
-    to_ignore = INDEX_COLUMNS | cs.ends_with("_x")
-    df = df.with_columns(pl.concat_str(~to_ignore, separator="-").alias("__group"))
+    group_cols = df.select(~(INDEX_COLUMNS | cs.ends_with("_x"))).collect_schema().names()
     return (
-        df.group_by("__group", maintain_order=True)
-        .agg((cs.ends_with("_x") & (~INDEX_COLUMNS)).unique(maintain_order=True).shuffle(seed).head(size))
-        .explode(x_cols, empty_as_null=False)
-        .join(df, on=["__group", *x_cols], how="left")
-        .select(cs.exclude("__group"))
+        df.group_by(group_cols, maintain_order=True)
+        .agg(pl.struct(x_cols).unique(maintain_order=True).shuffle(seed).head(size).alias("__group"))
+        .explode("__group", empty_as_null=False)
+        .unnest("__group")
+        .join(df, on=[*group_cols, *x_cols], how="left", maintain_order="left")
     )
 
 
@@ -49,7 +48,8 @@ class Subsampler:
     """Subsample the ABX :py:class:`.Task`.
 
     Each cell is limited to ``max_size_group`` items for A, B and X independently.
-    When using "across" conditions, each group of (A, B) is limited to ``max_x_across`` possible values for X.
+    When using "across" conditions, each group of (A, B) is limited to ``max_x_across`` observed
+    combinations of X condition values.
     Subsampling for one or more conditions can be disabled by setting the corresponding argument to ``None``.
 
     .. note::